requirements.txt
anilibria.py
dnspython
git+https://github.com/interactions-py/library.git@a2eaffdcf62ce6e0fdb1207be2a0d6e5aa98ba43
motor
pymongo
python-dotenv
